Remove ContactUsForm story from Storybook and move the component into communication-framework
